Update README.md
Having used this plugin (awesome by the way), I thought I'd improve the installation by adding 4. and I've put some additional information in 7-9 in the usage. Also put a bit of an intro in, so people can find it properly via search engines, as it took me ages to find this tool, and it does exactly what people need when they're in a situation where they've a couple of dozen/hundred URLs they need to get unindexed quickly, and doing it one by one would take hours.